backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
psycopg==3.2.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
gunicorn==23.0.0
djangorestframework-simplejwt==5.5.1 
pyjwt==2.10.1
redis==5.3.1